ipmi-sel: refactor date range parsing code
This is an attempt to refactor the bits in the date parsing code that is
trying to match the given datetime format among the supported ones.

The major change is to get rid of nested if statements during the
figuring out the input date format. Instead with the new approach we are
iterating over a constant array of supported datetime formats.

This patch also includes a bugfix. Before this patch, the extra
characters after valid date strings were accepted. For example a date
range such as the following was not causing any errors:

  $ ipmi-sel --date-range=01/01/2000foo-01/02/2000bar

This is due to the return value checking of the strptime() function that
is used. With this patch, the date strings are expected to match the
format exactly. So the aforementioned example will return an error now.

Signed-off-by: Mert Kirpici <mert.kirpici@canonical.com>